+++ resharper/resharper-unity/src/CSharp/Feature/Services/QuickFixes/CallGraph/PerformanceAnalysis/PerformanceDisableUtil.cs
@@ -1,7 +1,0 @@
-namespace JetBrains.ReSharper.Plugins.Unity.CSharp.Feature.Services.QuickFixes.CallGraph.PerformanceAnalysis
-{
-    public static class PerformanceDisableUtil
-    {
-        public const string MESSAGE = "Disable performance analysis for method";
-    }
-}