+++ resharper/resharper-unity/src/Rider/Highlightings/IconsProviders/RiderEventHandlerDetector.cs
@@ -54,14 +54,14 @@ namespace JetBrains.ReSharper.Plugins.Unity.Rider.Highlightings.IconsProviders
         }
 
         protected override void AddHighlighting(IHighlightingConsumer consumer, ICSharpDeclaration element, string text, string tooltip,
-                                                DaemonProcessKind kind)
+                                                IReadOnlyCallGraphContext context)
         {
-            var iconId = element.HasHotIcon(ContextProvider, SettingsStore.BoundSettingsStore, kind)
+            var iconId = element.HasHotIcon(ContextProvider, SettingsStore.BoundSettingsStore, context)
                 ? InsightUnityIcons.InsightHot.Id
                 : InsightUnityIcons.InsightUnity.Id;
 
             if (RiderIconProviderUtil.IsCodeVisionEnabled(SettingsStore.BoundSettingsStore, myCodeInsightProvider.ProviderId,
-                () => { base.AddHighlighting(consumer, element, text, tooltip, kind); }, out var useFallback))
+                () => { base.AddHighlighting(consumer, element, text, tooltip, context); }, out var useFallback))
             {
                 if (!useFallback)
                 {